chore: restructure token managers into a directory
