+++ ibm_cloud_sdk_core/authenticators/cp4d_authenticator.py
@@ -19,7 +19,7 @@ from typing import Dict, Optional
 from requests import Request
 
 from .authenticator import Authenticator
-from ..cp4d_token_manager import CP4DTokenManager
+from ..token_managers.cp4d_token_manager import CP4DTokenManager
 from ..utils import has_bad_first_or_last_char
