+++ ibm_cloud_sdk_core/token_managers/cp4d_token_manager.py
@@ -16,6 +16,7 @@
 
 import json
 from typing import Dict, Optional
+
 from .jwt_token_manager import JWTTokenManager
